# Cache `When` predicate