No need to re-compile the `Expression` into a `Func` each time `AppliesTo` is
called.
Made a simple benchmark to verify the improvement.

```c#
list.Should().BeEquivalentTo(list2, opt => opt
    .WithStrictOrdering()
    .Using<string>(ctx => ctx.Should().NotBeNull())
    .When(ctx => ctx.SelectedMemberPath == "Id")
);
```

Before:
```
|         Method |     N |         Mean |      Error |     StdDev |      Gen 0 |      Gen 1 | Gen 2 |    Allocated |
|--------------- |------ |-------------:|-----------:|-----------:|-----------:|-----------:|------:|-------------:|
| BeEquivalentTo |    10 |     2.685 ms |  0.0392 ms |  0.0367 ms |    74.2188 |    35.1563 |     - |    236.15 KB |
| BeEquivalentTo |   100 |    25.310 ms |  0.4918 ms |  0.4360 ms |   687.5000 |   343.7500 |     - |   2177.74 KB |
| BeEquivalentTo |  1000 |   241.684 ms |  4.7883 ms |  8.1309 ms |  7000.0000 |  3000.0000 |     - |  21595.78 KB |
| BeEquivalentTo |  5000 | 1,263.053 ms | 24.2829 ms | 33.2387 ms | 35000.0000 | 17000.0000 |     - | 107984.95 KB |
| BeEquivalentTo | 10000 | 2,532.558 ms | 50.3574 ms | 97.0216 ms | 69000.0000 | 32000.0000 |     - | 215943.08 KB |
```

After:
```
|         Method |     N |         Mean |       Error |      StdDev |      Gen 0 | Gen 1 | Gen 2 |    Allocated |
|--------------- |------ |-------------:|------------:|------------:|-----------:|------:|------:|-------------:|
| BeEquivalentTo |    10 |     259.4 us |     3.57 us |     2.79 us |    49.3164 |     - |     - |    151.77 KB |
| BeEquivalentTo |   100 |   1,369.3 us |    25.93 us |    26.63 us |   435.5469 |     - |     - |    1332.1 KB |
| BeEquivalentTo |  1000 |  12,348.7 us |   108.48 us |    84.70 us |  4281.2500 |     - |     - |  13138.82 KB |
| BeEquivalentTo |  5000 |  61,062.4 us | 1,198.94 us | 1,282.86 us | 21333.3333 |     - |     - |  65675.79 KB |
| BeEquivalentTo | 10000 | 125,624.8 us | 2,444.84 us | 2,910.41 us | 42000.0000 |     - |     - | 131347.25 KB |
```

10-20x faster
35-45% less allocated memory